providers/vmware: ensure it builds on unsupported architectures
